[WIP] Extract ble_qwr from estc_ble
